Moved operation in proper com/faacets/operation subfolders.
