+++ operation/src/main/scala/com/faacets/Operation.scala
