+++ operation/src/main/scala/com/faacets/operation/Affine.scala
