+++ operation/src/main/scala/com/faacets/operation/Decomposition.scala
