+++ operation/src/main/scala/com/faacets/operation/Extractor.scala
