+++ operation/src/main/scala/com/faacets/operation/Lifting.scala
