+++ operation/src/main/scala/com/faacets/operation/OperationNode.scala
