+++ operation/src/main/scala/com/faacets/operation/Redundant.scala
