+++ operation/src/main/scala/com/faacets/operation/SymmetricForms.scala
