+++ operation/src/main/scala/com/faacets/operation/affine/AffineGroup.scala
