+++ operation/src/main/scala/com/faacets/operation/affine/AffineParsable.scala
