+++ operation/src/main/scala/com/faacets/operation/affine/CorrAffineExtractor.scala
