+++ operation/src/main/scala/com/faacets/operation/affine/VecAffineAction.scala
