+++ operation/src/main/scala/com/faacets/operation/instances/all.scala
@@ -1,0 +1,5 @@
+package operation.instances
+
+trait AllInstances {
+
+}
