+++ operation/src/main/scala/com/faacets/operation/instances/package.scala
@@ -1,0 +1,9 @@
+package com.faacetsoperation
+
+package object instances {
+
+  object all extends AllInstances
+
+  object expr extends ExprInstances
+
+}
