+++ operation/src/main/scala/com/faacets/operation/lifting/Grouping.scala
