+++ operation/src/main/scala/com/faacets/operation/lifting/InputGrouping.scala
