+++ operation/src/main/scala/com/faacets/operation/lifting/LiftingExtractor.scala
