+++ operation/src/main/scala/com/faacets/operation/lifting/LiftingGroupoid.scala
