+++ operation/src/main/scala/com/faacets/operation/lifting/PartyGrouping.scala
