+++ operation/src/main/scala/com/faacets/operation/product/ExprTensorExtractor.scala
