+++ operation/src/main/scala/com/faacets/operation/product/Rank1.scala
