+++ operation/src/main/scala/com/faacets/operation/redundant/RedundantGroupoid.scala
