+++ operation/src/main/scala/com/faacets/operation/relabeling/SingleParty.scala
