+++ operation/src/main/scala/com/faacets/operation/reordering/ReorderingGroupoid.scala
