+++ operation/src/main/scala/com/faacets/operation/reordering/ScenarioReordering.scala
