+++ operation/src/main/scala/com/faacets/operation/reordering/VecReordering.scala
